IN-317 ref: took out a group of buttons
